refactor: split logging
